requirements.txt
chevron==0.14.0
Markdown==3.5.2
python-docx==1.1.0
PyYAML==6.0.1